+++ flask/.dockerignore
@@ -5,3 +5,4 @@ README.md
 *.pyd
 __pycache__
 .virtualenv
+env
