+++ flask/.gcloudignore
@@ -1,0 +1,42 @@
+
+# This file specifies files that are *not* uploaded to Google Cloud Platform
+# using gcloud. It follows the same syntax as .gitignore, with the addition of
+# "#!include" directives (which insert the entries of the given .gitignore-style
+# file at that point).
+#
+# For more information, run:
+#   $ gcloud topic gcloudignore
+#
+.gcloudignore
+# If you would like to upload your .git directory, .gitignore file or files
+# from your .gitignore file, remove the corresponding line
+# below:
+.git
+.gitignore
+
+# Python pycache:
+__pycache__/
+# Ignored by the build system
+/setup.cfg
+
+
+env
+env/
+
+.virtualenv
+**/.virtualenv
+
+node_modules/
+**/node_modules
+
+.dockerignore
+.env.default
+app.pyc
+db_og.py
+db.pyc
+Dockerfile
+Makefile
+script.py
+utils.pyc
+*.pyc
+**/*.pyc
