+++ react/.gcloudignore
@@ -1,0 +1,32 @@
+
+# This file specifies files that are *not* uploaded to Google Cloud Platform
+# using gcloud. It follows the same syntax as .gitignore, with the addition of
+# "#!include" directives (which insert the entries of the given .gitignore-style
+# file at that point).
+#
+# For more information, run:
+#   $ gcloud topic gcloudignore
+#
+.gcloudignore
+# If you would like to upload your .git directory, .gitignore file or files
+# from your .gitignore file, remove the corresponding line
+# below:
+.git
+.gitignore
+
+# Python pycache:
+__pycache__/
+# Ignored by the build system
+/setup.cfg
+
+**/.virtualenv
+flask/
+
+node_modules/
+**/node_modules
+Dockerfile
+Makefile
+.nvmrc
+.env.default
+.env.development
+.dockerignore
