+++ react/app.yaml
@@ -1,0 +1,13 @@
+env: standard
+service: tracing-react
+runtime: nodejs14
+
+handlers:
+  - url: /static
+    static_dir: build/static
+  - url: /(.*\.(json|ico|js))$
+    static_files: build/\1
+    upload: build/.*\.(json|ico|js)$
+  - url: .*
+    static_files: build/index.html
+    upload: build/index.html
